add t2 generator structure

+++ src/main/java/com/teammoeg/frostedheart/steamenergy/DebugHeaterTileEntity.java
@@ -57,7 +57,7 @@ public class DebugHeaterTileEntity extends IEBaseTileEntity implements HeatProvi
 
 
 	@Override
-	public float getTemperatureLevel() {
+	public int getTemperatureLevel() {
 		return 1;
 	}
