+++ src/main/java/com/teammoeg/frostedheart/steamenergy/HeatProvider.java
@@ -4,5 +4,5 @@ public interface HeatProvider extends EnergyNetworkProvider{
 	
 	float getMaxHeat();
 	float drainHeat(float value);
-	float getTemperatureLevel();
+	int getTemperatureLevel();
 }
